\name{BladTumor1}
\alias{BladTumor1}
\docType{data}
\title{
Converted Bladder Tumor Cancer Data
}
\description{
This data set is converted from \code{\link{BladTumor}} to have the format to be used in the function. One should convert the original data into this specified data structure and do the further analysis.
}
\usage{data(BladTumor1)}

\format{
  \tabular{ll}{
  t:\tab observation times matrix\cr
  x:\tab covariate matrix\cr
  z:\tab panel counts from non-overlapping intervals matrix\cr
  k:\tab the number of observations from all patients\cr
  K:\tab the largest number of observations\cr
  }
}

\seealso{
\code{\link{BladTumor}}
}

\keyword{datasets}
