\name{PCDSpline-package}
\alias{PCDSpline-package}
\alias{PCDSpline}
\docType{package}
\title{
A statistical package for regression analysis of panel count data under the non-homogeneous Poisson process models with and without frailty
}
\description{
This package allows for semiparametric regression analysis of panel count data under the non-homogeneous Poisson process models with and without frailty. The Gamma frailty model allows to account for the within-subject correlation. Monotone splines of Ramsay (1988) are used to estimate the unknown baseline mean function. Fitting the models to panel count data via EM algorithm.
}
\details{
\tabular{ll}{
Package: \tab PCDSpline\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Bin Yao <yaob@email.sc.edu> and Lianming Wang <wangl@stat.sc.edu>
}

